requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
idna==3.3
numpy==1.21.4
Pillow==8.4.0
requests==2.26.0
scrollphathd==1.3.0
smbus2==0.4.1
urllib3==1.26.7
